Add feature: New game from scratch
Added a New Game button below the game board. When clicked, it pauses
the game (if two bots are still playing), and returns to the home page,
allowing players to start a new game.

If a game has already been created, `startGame` in `gameSettings` won't
reinitializing `gameController`. Instead, it first resets round data
then executes the `newGame` function.

Implemented `newGame` in `gameController` to update game parameters,
reset the board, winner, and scores, and call `setPlayers` to recreate
players. The active and first players are also reset.